+++ lib/bmg/algebra/shortcuts.rb
@@ -9,6 +9,22 @@ module Bmg
         self.restrict(predicate)
       end
 
+      def prefix(prefix)
+        raise "Attrlist must be known to use `prefix`" unless self.type.knows_attrlist?
+        renaming = self.type.to_attrlist.each_with_object({}){|a,r|
+          r[a] = :"#{prefix}#{a}"
+        }
+        self.rename(renaming)
+      end
+
+      def suffix(suffix)
+        raise "Attrlist must be known to use `suffix`" unless self.type.knows_attrlist?
+        renaming = self.type.to_attrlist.each_with_object({}){|a,r|
+          r[a] = :"#{a}#{suffix}"
+        }
+        self.rename(renaming)
+      end
+
     end # module Shortcuts
   end # module Algebra
 end # module Bmg